merge

+++ hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/dao/problem/impl/ProblemEntityServiceImpl.java
@@ -4,7 +4,9 @@ import cn.hutool.core.bean.BeanUtil;
 import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.io.file.FileReader;
 import cn.hutool.core.io.file.FileWriter;
+import cn.hutool.core.lang.UUID;
 import cn.hutool.core.util.CharsetUtil;
+import cn.hutool.core.util.StrUtil;
 import cn.hutool.json.JSONArray;
 import cn.hutool.json.JSONObject;
 import cn.hutool.json.JSONUtil;
@@ -17,6 +19,7 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
 import org.springframework.util.DigestUtils;
 import org.springframework.util.StringUtils;
 import top.hcode.hoj.dao.judge.JudgeEntityService;
@@ -24,16 +27,17 @@ import top.hcode.hoj.dao.problem.*;
 import top.hcode.hoj.exception.ProblemIDRepeatException;
 import top.hcode.hoj.mapper.ProblemMapper;
 import top.hcode.hoj.pojo.bo.Pair_;
-import top.hcode.hoj.pojo.dto.ProblemDto;
+import top.hcode.hoj.pojo.dto.ProblemDTO;
 import top.hcode.hoj.pojo.entity.problem.*;
-import top.hcode.hoj.pojo.vo.ImportProblemVo;
-import top.hcode.hoj.pojo.vo.ProblemCountVo;
-import top.hcode.hoj.pojo.vo.ProblemVo;
+import top.hcode.hoj.pojo.vo.ImportProblemVO;
+import top.hcode.hoj.pojo.vo.ProblemCountVO;
+import top.hcode.hoj.pojo.vo.ProblemVO;
 import top.hcode.hoj.utils.Constants;
 
 import java.io.File;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 /**
@@ -71,25 +75,27 @@ public class ProblemEntityServiceImpl extends ServiceImpl<ProblemMapper, Problem
     @Autowired
     private CodeTemplateEntityService codeTemplateEntityService;
 
+    private final static Pattern EOL_PATTERN = Pattern.compile("[^\\S\\n]+(?=\\n)");
+
     @Override
-    public Page<ProblemVo> getProblemList(int limit, int currentPage, Long pid, String title, Integer difficulty,
+    public Page<ProblemVO> getProblemList(int limit, int currentPage, Long pid, String title, Integer difficulty,
                                           List<Long> tid, String oj) {
 
         //新建分页
-        Page<ProblemVo> page = new Page<>(currentPage, limit);
+        Page<ProblemVO> page = new Page<>(currentPage, limit);
         Integer tagListSize = null;
         if (tid != null) {
             tid = tid.stream().distinct().collect(Collectors.toList());
             tagListSize = tid.size();
         }
 
-        List<ProblemVo> problemList = problemMapper.getProblemList(page, pid, title, difficulty, tid, tagListSize, oj);
+        List<ProblemVO> problemList = problemMapper.getProblemList(page, pid, title, difficulty, tid, tagListSize, oj);
 
         if (problemList.size() > 0) {
-            List<Long> pidList = problemList.stream().map(ProblemVo::getPid).collect(Collectors.toList());
-            List<ProblemCountVo> problemListCount = judgeEntityService.getProblemListCount(pidList);
-            for (ProblemVo problemVo : problemList) {
-                for (ProblemCountVo problemCountVo : problemListCount) {
+            List<Long> pidList = problemList.stream().map(ProblemVO::getPid).collect(Collectors.toList());
+            List<ProblemCountVO> problemListCount = judgeEntityService.getProblemListCount(pidList);
+            for (ProblemVO problemVo : problemList) {
+                for (ProblemCountVO problemCountVo : problemListCount) {
                     if (problemVo.getPid().equals(problemCountVo.getPid())) {
                         problemVo.setProblemCountVo(problemCountVo);
                         break;
@@ -103,7 +109,7 @@ public class ProblemEntityServiceImpl extends ServiceImpl<ProblemMapper, Problem
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public boolean adminUpdateProblem(ProblemDto problemDto) {
+    public boolean adminUpdateProblem(ProblemDTO problemDto) {
 
         Problem problem = problemDto.getProblem();
         if (Constants.JudgeMode.DEFAULT.getMode().equals(problemDto.getJudgeMode())) {
@@ -385,7 +391,7 @@ public class ProblemEntityServiceImpl extends ServiceImpl<ProblemMapper, Problem
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public boolean adminAddProblem(ProblemDto problemDto) {
+    public boolean adminAddProblem(ProblemDTO problemDto) {
 
         Problem problem = problemDto.getProblem();
 
@@ -394,26 +400,39 @@ public class ProblemEntityServiceImpl extends ServiceImpl<ProblemMapper, Problem
                     .setSpjCode(null);
         }
 
-        /**
-         *  problem_id唯一性检查
-         */
-        String problemId = problem.getProblemId().toUpperCase();
-        QueryWrapper<Problem> problemQueryWrapper = new QueryWrapper<>();
-        problemQueryWrapper.eq("problem_id", problemId);
-        int existedProblem = problemMapper.selectCount(problemQueryWrapper);
-        if (existedProblem > 0) {
-            throw new ProblemIDRepeatException("The problem_id [" + problemId + "] already exists. Do not reuse it!");
-        }
-
-
         // 设置测试样例的版本号
         problem.setCaseVersion(String.valueOf(System.currentTimeMillis()));
         if (problem.getIsGroup() == null) {
             problem.setIsGroup(false);
         }
-        problem.setProblemId(problemId);
-        boolean addProblemResult = problemMapper.insert(problem) == 1;
-        long pid = problem.getId();
+
+        // 如果没有提供problemId,则或者生成 P1000之类的，以problem表的id作为数字
+        if (problem.getProblemId() == null) {
+            problem.setProblemId(UUID.fastUUID().toString());
+            problemMapper.insert(problem);
+
+            UpdateWrapper<Problem> problemUpdateWrapper = new UpdateWrapper<>();
+            problemUpdateWrapper.set("problem_id", "P" + problem.getId());
+            problemUpdateWrapper.eq("id", problem.getId());
+            problemMapper.update(null, problemUpdateWrapper);
+            problem.setProblemId("P" + problem.getId());
+        } else {
+            // problem_id唯一性检查
+            String problemId = problem.getProblemId().toUpperCase();
+            QueryWrapper<Problem> problemQueryWrapper = new QueryWrapper<>();
+            problemQueryWrapper.eq("problem_id", problemId);
+            int existedProblem = problemMapper.selectCount(problemQueryWrapper);
+            if (existedProblem > 0) {
+                throw new ProblemIDRepeatException("The problem_id [" + problemId + "] already exists. Do not reuse it!");
+            }
+            problem.setProblemId(problemId);
+            problemMapper.insert(problem);
+        }
+        Long pid = problem.getId();
+        if (pid == null) {
+            throw new ProblemIDRepeatException("The problem with problem_id [" + problem.getProblemId() + "] insert failed!");
+        }
+
         // 为新的题目添加对应的language
         List<ProblemLanguage> problemLanguageList = new LinkedList<>();
         for (Language language : problemDto.getLanguages()) {
@@ -514,7 +533,7 @@ public class ProblemEntityServiceImpl extends ServiceImpl<ProblemMapper, Problem
         }
 
 
-        if (addProblemResult && addCasesToProblemResult && addLangToProblemResult
+        if (addCasesToProblemResult && addLangToProblemResult
                 && addTagsToProblemResult && addProblemCodeTemplate) {
             return true;
         } else {
@@ -534,10 +553,15 @@ public class ProblemEntityServiceImpl extends ServiceImpl<ProblemMapper, Problem
         String testCasesDir = Constants.File.TESTCASE_BASE_FOLDER.getPath() + File.separator + "problem_" + problemId;
 
         // 将之前的临时文件夹里面的评测文件全部复制到指定文件夹(覆盖)
+
         if (!StringUtils.isEmpty(tmpTestcaseDir)) {
             FileUtil.clean(testCasesDir);
-            FileUtil.copyFilesFromDir(new File(tmpTestcaseDir), new File(testCasesDir), true);
+            File testCasesDirFile = new File(testCasesDir);
+            FileUtil.copyFilesFromDir(new File(tmpTestcaseDir), testCasesDirFile, true);
         }
+
+        List<String> listFileNames = FileUtil.listFileNames(testCasesDir);
+
         if (StringUtils.isEmpty(judgeCaseMode)) {
             judgeCaseMode = Constants.JudgeCaseMode.DEFAULT.getMode();
         }
@@ -556,23 +580,37 @@ public class ProblemEntityServiceImpl extends ServiceImpl<ProblemMapper, Problem
                     || judgeCaseMode.equals(Constants.JudgeCaseMode.SUBTASK_LOWEST.getMode())) {
                 jsonObject.set("groupNum", problemCase.getGroupNum());
             }
+
             jsonObject.set("score", problemCase.getScore());
             jsonObject.set("inputName", problemCase.getInput());
             jsonObject.set("outputName", problemCase.getOutput());
 
+            listFileNames.remove(problemCase.getInput());
+            listFileNames.remove(problemCase.getOutput());
+
             // 读取输入文件
             FileReader inputFile = new FileReader(testCasesDir + File.separator + problemCase.getInput(), CharsetUtil.UTF_8);
-            String input = inputFile.readString().replaceAll("\r\n", "\n");
+            String input = inputFile.readString()
+                    .replaceAll("\r\n", "\n") // 避免window系统的换行问题
+                    .replaceAll("\r", "\n"); // 避免mac系统的换行问题
 
             FileWriter inputFileWriter = new FileWriter(testCasesDir + File.separator + problemCase.getInput(), CharsetUtil.UTF_8);
             inputFileWriter.write(input);
 
             // 读取输出文件
-            FileReader outputFile = new FileReader(testCasesDir + File.separator + problemCase.getOutput(), CharsetUtil.UTF_8);
-            String output = outputFile.readString().replaceAll("\r\n", "\n");
-
-            FileWriter outFileWriter = new FileWriter(testCasesDir + File.separator + problemCase.getOutput(), CharsetUtil.UTF_8);
-            outFileWriter.write(output);
+            String output = "";
+            String outputFilePath = testCasesDir + File.separator + problemCase.getOutput();
+            if (FileUtil.exist(outputFilePath)) {
+                FileReader outputFile = new FileReader(outputFilePath, CharsetUtil.UTF_8);
+                output = outputFile.readString()
+                        .replaceAll("\r\n", "\n") // 避免window系统的换行问题
+                        .replaceAll("\r", "\n"); // 避免mac系统的换行问题
+                FileWriter outFileWriter = new FileWriter(testCasesDir + File.separator + problemCase.getOutput(), CharsetUtil.UTF_8);
+                outFileWriter.write(output);
+            } else {
+                FileWriter fileWriter = new FileWriter(outputFilePath);
+                fileWriter.write("");
+            }
 
             // spj和interactive是根据特判程序输出判断结果，所以无需初始化测试数据
             if (Constants.JudgeMode.DEFAULT.getMode().equals(judgeMode)) {
@@ -596,6 +634,13 @@ public class ProblemEntityServiceImpl extends ServiceImpl<ProblemMapper, Problem
         infoFile.write(JSONUtil.toJsonStr(result));
         // 删除临时上传文件夹
         FileUtil.del(tmpTestcaseDir);
+        // 删除非测试数据的文件
+        listFileNames.remove("info");
+        if (!CollectionUtils.isEmpty(listFileNames)) {
+            for (String filename : listFileNames) {
+                FileUtil.del(testCasesDir + File.separator + filename);
+            }
+        }
     }
 
 
@@ -633,13 +678,21 @@ public class ProblemEntityServiceImpl extends ServiceImpl<ProblemMapper, Problem
             // 生成对应文件
             FileWriter infileWriter = new FileWriter(testCasesDir + "/" + inputName, CharsetUtil.UTF_8);
             // 将该测试数据的输入写入到文件
-            String inputData = problemCaseList.get(index).getInput().replaceAll("\r\n", "\n");
+            String inputData = problemCaseList
+                    .get(index)
+                    .getInput()
+                    .replaceAll("\r\n", "\n") // 避免window系统的换行问题
+                    .replaceAll("\r", "\n"); // 避免mac系统的换行问题
             infileWriter.write(inputData);
 
             String outputName = (index + 1) + ".out";
             jsonObject.set("outputName", outputName);
             // 生成对应文件
-            String outputData = problemCaseList.get(index).getOutput().replaceAll("\r\n", "\n");
+            String outputData = problemCaseList
+                    .get(index)
+                    .getOutput()
+                    .replaceAll("\r\n", "\n") // 避免window系统的换行问题
+                    .replaceAll("\r", "\n"); // 避免mac系统的换行问题
             FileWriter outFile = new FileWriter(testCasesDir + "/" + outputName, CharsetUtil.UTF_8);
             outFile.write(outputData);
 
@@ -668,10 +721,10 @@ public class ProblemEntityServiceImpl extends ServiceImpl<ProblemMapper, Problem
 
     @Override
     @SuppressWarnings("All")
-    public ImportProblemVo buildExportProblem(Long pid, List<HashMap<String, Object>> problemCaseList,
+    public ImportProblemVO buildExportProblem(Long pid, List<HashMap<String, Object>> problemCaseList,
                                               HashMap<Long, String> languageMap, HashMap<Long, String> tagMap) {
         // 导出相当于导入
-        ImportProblemVo importProblemVo = new ImportProblemVo();
+        ImportProblemVO importProblemVo = new ImportProblemVO();
         Problem problem = problemMapper.selectById(pid);
         problem.setCaseVersion(null)
                 .setGmtCreate(null)
@@ -724,12 +777,7 @@ public class ProblemEntityServiceImpl extends ServiceImpl<ProblemMapper, Problem
     // 去除每行末尾的空白符
     public static String rtrim(String value) {
         if (value == null) return null;
-        StringBuilder sb = new StringBuilder();
-        String[] strArr = value.split("\n");
-        for (String str : strArr) {
-            sb.append(str.replaceAll("\\s+$", "")).append("\n");
-        }
-        return sb.toString().replaceAll("\\s+$", "");
+        return EOL_PATTERN.matcher(StrUtil.trimEnd(value)).replaceAll("");
     }
